Small changes related to BF16 fusion_gru and fusion_lstm
