+++ python/paddle/fluid/tests/unittests/mkldnn/test_fusion_gru_int8_mkldnn_op.py
@@ -35,6 +35,7 @@ class TestFusionGRUINT8MKLDNNOp(OpTest):
         self.act_gate = 'sigmoid'
         self.origin_mode = True
         self.use_mkldnn = True
+        self.mkldnn_data_type = "int8"
         self.force_fp32_output = True
         self.error_margin = 1e-5
         self.set_confs()
@@ -115,6 +116,7 @@ class TestFusionGRUINT8MKLDNNOp(OpTest):
             'is_reverse': self.is_reverse,
             'origin_mode': self.origin_mode,
             'use_mkldnn': self.use_mkldnn,
+            'mkldnn_data_type': self.mkldnn_data_type,
             'force_fp32_output': self.force_fp32_output,
             'Scale_data': scale_data,
             'Shift_data': shift_data,
