+++ python/paddle/fluid/tests/unittests/mkldnn/test_fusion_lstm_int8_mkldnn_op.py
@@ -34,6 +34,7 @@ class TestFusionLSTMINT8MKLDNNOp(OpTest):
         self.act_cand = 'tanh'
         self.use_peepholes = False  # LSTM u8 doesn't support peepholes
         self.use_mkldnn = True
+        self.mkldnn_data_type = "int8"
         self.force_fp32_output = False
         self.error_margin = 1e-5
         self.set_confs()
@@ -117,6 +118,7 @@ class TestFusionLSTMINT8MKLDNNOp(OpTest):
             'is_reverse': self.is_reverse,
             'use_peepholes': self.use_peepholes,
             'use_mkldnn': self.use_mkldnn,
+            'mkldnn_data_type': self.mkldnn_data_type,
             'force_fp32_output': self.force_fp32_output,
             'Scale_data': scale_data,
             'Shift_data': shift_data,
